requirements.txt
requests
feedparser
beautifulsoup4
sqlalchemy
pyyaml
python-dotenv
apscheduler
openai
